+++ CHANGELOG.md
@@ -25,6 +25,17 @@ should clarify that these functions are indeed suitable for inclusion in V2 scri
    As a result, some existing tests for LPE have been temporarily removed. [Issue #53](https://github.com/Liqwid-Labs/liqwid-plutarch-extra/issues/53)
    has been opened to port these tests to PQ2.0
 
+## 2.0.1 -- 2022-08-11
+
+### Added
+
+- `pjust` and `pnothing` for easier construction of `PJust` value.
+- `pmaybe` which has the same semantics as `Data.Maybe.maybe`.
+
+### Changed
+
+- Rename the original `pamybe` to `pfromMaybe`.
+
 ## 2.0.0 -- 2022-08-02
 
 ### Added